requirements.txt
asgiref==3.7.2
Django==4.2.7
sqlparse==0.4.4
djangorestframework==3.14.0
pytz==2023.3.post1
PyYAML==6.0.1

tzdata==2023.3
